Fix incorrect buffer length in transmision
In the communication between client and server the data is trimmed.
The error is caused because an incorect length is specified when
the DatagramPacket is created. The code before fix used the length
of the data string. This did not match the length of the byte array
which is actually transmitted. This has been fixed by using the
correct length in the DatagramPacket constructor.

Fixes: https://github.com/a-rmz/JEasyCrypto/issues/7