Fix client close exception
Exception thrown because the socket is called from the main file,
while the ResponseReader is blocked on the socket reader.
Because it is a normal process in java thread,
the exit should be handled more gracefully,
by exiting without printing an error message.
So the fixes add flags to the exception,
so when the interrupt function is called and
it catches the exception,
it should exit without printing an error message.

fixes: https://github.com/a-rmz/JEasyCrypto/issues/23#issue-264173987