Extract ui logic from CryptoClient
This commit extracts the ui logic from CryptoClient class into
ConsoleLauncher class. The purpose is to isolate the user interface
from the network communication so that the console user interface
can be easily replaced with a GUI. The new implementation of
CryptoClient handles only the network communication and notifies
the user interface when a new message is available.