Add gui to the system
A graphical user interface developed in javafx was created
for the system. It contains two main screens: a start up
screen, where the user introduces the server address, and
a main scree where the user can perform the encrypt and
decrypt operations. A change was made to the constructor
of CryptoClient so that it would no longer receive the
observer object. This object has to be registered using the
method setObserver after the object is initialized. The
change was need to accomodate with the design of the gui
classes.

Fixes: https://github.com/a-rmz/JEasyCrypto/issues/16